# docs/swagger.yaml
definitions:
  domain.AccountType:
    enum:
    - bank
    - cash
    - credit_card
    - investment
    - other
    type: string
    x-enum-varnames:
    - AccountTypeBank
    - AccountTypeCash
    - AccountTypeCreditCard
    - AccountTypeInvestment
    - AccountTypeOther
  dto.AccountResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      currency:
        type: string
      icon:
        type: string
      id:
        type: string
      initial_balance:
        type: number
      name:
        type: string
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.AccountType'
      updated_at:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.CategoryResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      parent_category_id:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.CreateAccountRequest:
    properties:
      color:
        type: string
      currency:
        type: string
      icon:
        type: string
      initial_balance:
        type: number
      name:
        type: string
      type:
        $ref: '#/definitions/domain.AccountType'
    required:
    - name
    - type
    type: object
  dto.CreateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        type: string
      parent_category_id:
        type: string
    required:
    - name
    type: object
  dto.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.CreateTenantRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  dto.TagResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deactivated_at:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.TenantResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.UpdateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        type: string
      parent_category_id:
        type: string
    required:
    - name
    type: object
  dto.UpdateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  dto.User:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.UserTenantResponse:
    properties:
      name:
        type: string
      tenant_id:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Financial tracking API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FinTrack API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: list accounts by tenant ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AccountResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: create a new account
      parameters:
      - description: Create account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Create an account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: get account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Get an account
      tags:
      - accounts
  /auth/login:
    post:
      consumes:
      - multipart/form-data
      description: Login with username and password
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with full name, email and password
      parameters:
      - description: Register User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /categories:
    get:
      description: Get all categories for the authenticated user's tenant
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category for the authenticated user's tenant
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Create Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Create category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Soft Delete a category
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Delete category
      tags:
      - categories
    get:
      description: Get a category by its ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Get category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Update category
      tags:
      - categories
  /tags:
    get:
      description: Get all tags for the authenticated user's tenant
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TagResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: List tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag for the authenticated user's tenant
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Create Tag Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Create tag
      tags:
      - tags
  /tags/{id}:
    delete:
      description: Soft Delete a tag
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Delete tag
      tags:
      - tags
    get:
      description: Get a tag by its ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Get tag
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Update an existing tag
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Tag Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Update tag
      tags:
      - tags
  /tenants:
    post:
      consumes:
      - application/json
      description: Create a new tenant (workspace) and links the creator to it.
      parameters:
      - description: Create tenant
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Create a new tenant
      tags:
      - tenants
  /users/profile:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: Update user profile
      tags:
      - users
  /users/tenants:
    get:
      description: Get the tenants of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserTenantResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AuthPassword: []
      summary: List user tenants
      tags:
      - users
securityDefinitions:
  AuthPassword:
    flow: password
    tokenUrl: /auth/login
    type: oauth2
swagger: "2.0"